orderbot/requirements.txt
python-telegram-bot[webhooks]==20.8
python-dotenv==1.0.1
gspread==6.0.2
schedule==1.2.1
google-auth==2.27.0
aiohttp==3.9.3
pytz==2024.1